FRWorker/requirements.txt
# Face recognition worker & web app (Python 3.10–3.13)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
huggingface_hub>=0.16.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.24.0
Flask>=3.0.0
Werkzeug>=3.0.0
omegaconf>=2.3.0
fvcore>=0.1.5
lancedb>=0.29.0
pyarrow>=14.0.0
pandas>=2.0.0
